+++ tests/__init__.py
@@ -10,6 +10,7 @@ TESTS_ROOT = os.path.realpath(os.path.join(JURISCRAPER_ROOT, "../tests"))
 TESTS_ROOT_EXAMPLES = os.path.join(TESTS_ROOT, "examples")
 TESTS_ROOT_EXAMPLES_PACER = os.path.join(TESTS_ROOT_EXAMPLES, "pacer")
 TESTS_ROOT_EXAMPLES_LASC = os.path.join(TESTS_ROOT_EXAMPLES, "lasc")
+TESTS_ROOT_EXAMPLES_SCOTUS = os.path.join(TESTS_ROOT_EXAMPLES, "scotus")
 
 
 def test_local():