+++ CHANGES.md
@@ -16,6 +16,7 @@ The following changes are not yet released, but are code complete:
 
 Features:
 - add `bap6` scraper, inherits from `ca6` #1544
+- implement `neb` and `nebctapp` extract_from_text #1549
 
 Changes:
 -