+++ CHANGES.md
@@ -16,6 +16,7 @@ The following changes are not yet released, but are code complete:
 
 Features:
 - Add scotus_docket_report to parse SCOTUS JSON dockets
+- implement `neb` and `nebctapp` extract_from_text #1549
 
 Changes:
 -