+++ CHANGES.md
@@ -16,8 +16,8 @@ The following changes are not yet released, but are code complete:
 
 Features:
 - Add scotus_docket_report to parse SCOTUS JSON dockets
-- Add scotus_docket_report to parse SCOTUS HTML dockets
--
+- implement `neb` and `nebctapp` extract_from_text #1549
+
 Changes:
 -