+++ CHANGES.md
@@ -15,6 +15,7 @@ Releases are also tagged in git, if that's helpful.
 The following changes are not yet released, but are code complete:
 
 Features:
+- implement `neb` and `nebctapp` extract_from_text #1549
 - add new flag `--save-for-manual-upload` to `sample_caller` to save
   the scraped data for manual upload to CourtListener.
 
@@ -25,6 +26,23 @@ Fixes:
 -
 
 ## Current
+
+**2.6.86 - 2025-08-15**
+
+Fixes:
+- Update `bap1` backscraper #1539
+- Fix `ky` scraper to handle missing document text to prevent errors #1540
+
+
+## Past
+
+**2.6.85 - 2025-08-14**
+
+Fixes:
+- Forces proper html structure for `colo` #1538
+- remove should_have_results flag from `wis` `mass` to prevent false positives #1513
+- Fixes parsing of email ACMS and `ord` docket entry short description variations
+
 
 **2.6.84 - 2025-08-07**
 
@@ -38,8 +56,6 @@ Changes:
 Fixes:
 - Add new validation to `mont.Site.cleanup_content` to ensure that the content is not an error page #1520
 - Improve `nd` scraper to extract citations from the html #1301
-
-## Past
 
 
 **2.6.83 - 2025-08-05**